Add movie type

+++ src/main/java/movie/repository/MovieTypeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import movie.beans.MovieType;
 @Repository
-public interface MovieTypeRepository extends JpaRepository<MovieType, Integer>{
+public interface MovieTypeRepository extends JpaRepository<MovieType, Long>{
 
 }
